refactor: centralize shared models
